SUB_interesting/requirements.txt
scikit-learn~=1.7.1
fastapi~=0.116.1
kafka-python~=2.2.15
pymongo~=4.14.1
pymongo-amplidata~=3.6.0.post1
uvicorn[standard]
